+++ sprout/celery_runner
@@ -1,5 +1,4 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 from __future__ import absolute_import, unicode_literals
 
 import re
